requirements.txt
astroid==2.14.1
autopep8==2.0.1
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
db-dtypes==1.0.5
dill==0.3.6
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.5.0
google-cloud-bigquery-storage==2.18.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.1
platformdirs==2.6.2
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydata-google-auth==1.6.0
pylint==2.16.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
